Tests not using run in separate process for now.
Trying to run tests in a separate process doesn't seem to work well.
When using --process-isolation on the command line, it is difficult to
determine the proper place to include files as an include file which
should make a class available at the top of the file still has an error
of that class missing within the test method.  Other issue also exist,
so just run all tests directly in the same process.