Class loader can return false if class file not found.
The change to the class loader allows a namespace to be extended.  This
can be useful if for instance there is a need to develop a custom
database or cache driver. The cache manager expects the driver class to
be mrbavii\Framework\Cache\Driver_<name>.  By allowing the namespace to
be extended, then a custom driver could be created outside of the source
tree of the framework, and a classloader registered so that it would be
loaded correctly.