Added admin key functions.
The admin key is not the password for a user on the board.  It is the
password used when performing functions that are not associated with a
user on the board, such as installing or ugrading the board software.
Neither of these actions actually log a user on the board, so the admin
key is a password that is required to continue with those actions.