Remove initial checks from App::execute
To allow for more flexibility in derived apps, App::execute no longer
performs the initial checks to ensure that there is path information,
that it doesn't end with a "/", and that the path components are secure.
This functionality can be placed in the beginning of the dispatcher file
and coded as needed for the application.