Templates now support relative paths.
Calling a template from within another template now supports relative
name calling.  This relative name is to the template name not any
actual file system path.  If the first template called is /admin/install
and it calls a template named "update", then the normalized template
would be "/admin/update".  If  template starts with "/" then it is
treated as absolute.  "." and ".." are handled as expected.